Reorganize revisions files
