+++ revisions/model.js
@@ -2,7 +2,7 @@ const hasha = require('hasha')
 const {omit} = require('lodash')
 const createError = require('http-errors')
 const {applyValidateBAL} = require('./validate-bal')
-const mongo = require('./util/mongo')
+const mongo = require('../lib/util/mongo')
 
 async function createRevision({context, codeCommune, client}) {
   const now = new Date()
